Move `CategoryWidgetProps` out of writeback types
