+++ frontend/src/metabase/writeback/internalTypes.ts
@@ -1,0 +1,11 @@
+import Field from "metabase-lib/lib/metadata/Field";
+
+export interface CategoryWidgetProps {
+  field: {
+    value: string;
+    onChange: (value: string) => void;
+  };
+  formField: {
+    fieldInstance: Field;
+  };
+}
